fix compile error due to CDT API change

+++ rdt/org.eclipse.ptp.rdt.ui/src/org/eclipse/ptp/internal/rdt/ui/callhierarchy/CHLabelProvider.java
@@ -85,7 +85,7 @@ public class CHLabelProvider extends LabelProvider implements IColorProvider {
             if (decl != null) {
             	String label;
             	if (node.isMultiDef()) {
-            		int options= fCLabelProvider.getTextFlags();
+            		long options= fCLabelProvider.getTextFlags();
             		fCLabelProvider.setTextFlags(LABEL_OPTIONS_SIMPLE);
             		label= fCLabelProvider.getText(decl);
             		fCLabelProvider.setTextFlags(options);
